Move useMediaSessions and useMediaControls into player provider
