+++ lib/hooks/use-player.ts
@@ -18,7 +18,7 @@ type PlayerContextProps = PlayerState & {
   nextEpisode: () => void
   prevEpisode: () => void
   toggleAudio: () => void
-  play: (trackId: TEpisode['id']) => void
+  play: (track: TEpisode) => void
   pause: () => void
   seekForward: (time: number) => void
   replay: (time: number) => void
